source/servers/cloudtrail-query-python/pyproject.toml
[project]
name = "cloudtrail-query-python"
version = "0.1.0"
description = "MCP server for querying AWS CloudTrail events"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.37.33",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.28.0",
    "starlette>=0.36.3",
]

[tool.ruff]
line-length = 99
extend-include = ["*.ipynb"]
exclude = [
    ".venv",
    "**/__pycache__",
    "**/node_modules",
    "**/dist",
    "**/build",
    "**/env",
    "**/.ruff_cache",
    "**/.venv",
    "**/.ipynb_checkpoints"
]
force-exclude = true

[tool.ruff.lint]
exclude = ["__init__.py"]
select = ["C", "D", "E", "F", "I", "W"]
ignore = ["C901", "E501", "E741", "F402", "F823", "D100", "D106"]

[tool.ruff.lint.isort]
lines-after-imports = 2
no-sections = true

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["F704"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
